[differential fuzzing][evaluation] cleaning and fixing evaluation scripts

+++ evaluation-ddyf/reproducing_cve.sh
@@ -1,35 +1,37 @@
 #!/bin/bash
 set -euo pipefail
 
-SEQ=0
+START=${1:-1}
+END=${2:-50}
+START_CORE=${3:-0}
+INIT=${4:-1}
 
 
 export LIBAFL_EDGES_MAP_SIZE=262144
 
-echo 'Cleaning previous data'
-cargo clean
-rm -rf objectives seeds corpus experiments
+if [ $INIT -eq 1 ]
+then
+    echo 'Cleaning previous data'
+    cargo clean
+    rm -rf objectives seeds corpus experiments
 
 
-echo 'Building fuzzer'
-./tools/mk_vendor make wolfssl:wolfssl510
-./tools/mk_vendor make openssl:openssl340
-cargo build --release --bin tlspuffin --features cputs
+    echo 'Building fuzzer'
+    ./tools/mk_vendor make wolfssl:wolfssl510
+    ./tools/mk_vendor make openssl:openssl340
+    cargo build --release --bin tlspuffin --features cputs
 
-echo 'Generate seeds for diff fuzzing'
-./target/release/tlspuffin seed --differential
-
-
+    echo 'Generate seeds for diff fuzzing'
+    ./target/release/tlspuffin seed --differential
+fi
 
 TIMEOUT='5h'
-START=$((1+10*$SEQ))
-END=$((10+10*$SEQ))
+CORE_PER_EXP=8
 
-FIRST_CORE=24
-START_CORE=$(($FIRST_CORE + 8*SEQ))
-END_CORE=$(($FIRST_CORE + 8 + 8*SEQ - 1))
+
+END_CORE=$(($START_CORE + CORE_PER_EXP - 1))
 CORES="$START_CORE-$END_CORE"
-PORT=$((10000 + $SEQ))
+PORT=$((10000 + $START_CORE))
 
 echo "Running campaigns $START to $END on cores $CORES using port $PORT"
